pyproject.toml
[build-system]
requires = ['setuptools>=61.0']
build-backend = 'setuptools.build_meta'

[project]
name = 'ts-scan'
version = '1.1.0'
authors = [
    { name='EACG GmbH' }
]
description = 'TrustSource Packages Scanner'
readme = 'README.md'
requires-python = '>=3.10'
classifiers = []
keywords = ['dependencies', 'modules', 'compliance', 'TrustSource']
license = { text = 'Apache-2.0'}

dependencies = [
    "defusedxml~=0.7.1",
    "semantic_version~=2.10.0",
    "build~=1.2.2",
    "importlib-metadata~=8.6.0",
    "requests~=2.32.3",
    "ts-deepscan~=2.3.0",
    "spdx-tools>=0.8.2",
    "click==8.1.7",
    "click-params~=0.5.0",
    "license-expression~=30.4.0",
    "packageurl-python~=0.16.0",
    "wasabi~=1.1.3",
    "dataclasses-json~=0.6.7",
    "cyclonedx-python-lib~=8.5.0",
    "shippinglabel~=2.1.0",
    "tqdm~=4.67.0"
]

[project.urls]
Homepage = 'https://github.com/trustsource/ts-scan.git'

[project.scripts]
ts-scan = "ts_scan.cli:start"

[tool.setuptools.packages.find]
where = ['src']

# [tool.setuptools.package-data]
# ts_scan.analyse.gitignore = ['ts_scan/analyse/gitignore/*.gitignore']
